drivers: thermal: lmh-dcvs: Register the cooling device in order
The CPU cooling device registration is scheduled for later execution
during probe. Due to this, the CPU cooling devices are not created in
the same order as CPU logical numbers. This could result in unexpected
behavior for thermal-engine.

Modify the driver to acquire a lock and ensure that the cooling devices
are always registered in the same order as CPU logical numbers.

Change-Id: I67dbfa8e386480fc437425386855faf2f6ebdd5d
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
Signed-off-by: KanonifyX <kanonify01@gmail.com>