input: ts: mxt640u: Refactor firmware requesting and loading
It's definitely not a good idea to totally block kernel init if
the firmware doesn't get found for any reason (FS not mounted yet
or fw unavailable or any other kind of issue), because that'd make
it to just lock up forever.

Refactor the firmware loading mechanism to be asynchronous: that
won't block the kernel while trying to load the firmware at device
probe time and will also get the boot times back to normal.

Change-Id: I7e37a642ee5278609345f5d8546a5954f187ce26
Signed-off-by: kanonifyX <kanonify01@gmail.com>