arm64: dts: qcom: sdm845: Allow cpuidle sysfs properties to be modified
Normally, cpuidle exposes read-only properties in sysfs such as
current_governor_ro but does not allow them to be changed. While the
ability to tune these properties is supposed to be reserved for
developer testing purposes, I don't see any disadvantages of it after
looking through the cpuidle code.

Enable cpuidle_sysfs_switch to allow these properties to be modified.

Change-Id: I435971fb24fa97e707745856137ad771fd9019f5
Signed-off-by: kdrag0n <dragon@khronodragon.com>